Move WallpaperUtils in WallpaperForwarder
